4.0.0 Update: Implement generic DAL save / load functionality
Everything is different now. Had to rework large chunks of ModelFactory
and rethink SabreFacade yet again.

This update introduces the IDataMapper interface, used by
ModelFactory when saving or loading. Saving or loading is done via
some (as yet largely untested) SQL generators that iterate through
the given IDataMapper implementation, using K_NAME as column names.
IDataMapper is just a shim at this stage -- I just moved some decls
out of IBabylonModel. The implementation is still done by BabylonModel
under the hood. But it provides a possible extension point if we ever
want to implement DataMapper classes that are independent of the
BabylonModels (probably a good idea but lots of work).

This update also does away with the notion I originally had of using
SabreFacade to cache instances of VCalendar / VEvent and provide
lookups for them, because the tree-like structure of Sabre VObjects
means they maintain references to one another and so caching them was
getting really messy whenver their corresponding IBabylonModel objects
mutated. Instead, I have introduced the system of using SabreFacade as
a source of copy-constructors from ICalendarModel / IEventModel into
VCalendar /  VEvent objects. ModelFactory implements methods that do
the reverse. This is slower, but deterministic!

There are also a few updates to some of the exception classes, and to
a few other things that I can't remember. Basically, I have been bad
and forgetting to make regular commits.